gpkg-to-pgdump: Disable trigger on push to master
Essentially, this limits the effective trigger to tagged releases only.

This to prevent the gpkg-to-pgdump workflow from running too often
in close proximity, which often led to race conditions when one or
more runs try to upload to S3 or B2 bucket at the same time,
leading to failure.

As the use of release assets on GitHub, which is tied to a tagged
release (e.g. v1.1), seems to be the best way forward for dissemination
of files such as opendrr-boundaries.dump, removing the trigger on push
to master branch seems to make better sense too.